requirements.txt
absl-py==0.5.0
astor==0.7.1
backports.functools-lru-cache==1.5
backports.weakref==1.0.post1
cloudpickle==0.6.0
cycler==0.10.0
dask==0.19.4
decorator==4.3.0
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
imutils==0.5.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==3.0.1
matplotlib==2.2.3
mock==2.0.0
networkx==2.2
numpy==1.15.2
opencv-python==3.4.3.18
pathlib==1.0.1
pbr==4.3.0
Pillow==5.3.0
protobuf==3.6.1
pyparsing==2.2.2
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
scikit-image==0.14.1
scipy==1.1.0
six==1.11.0
subprocess32==3.5.3
tensorboard==1.11.0
tensorflow==1.11.0
termcolor==1.1.0
toolz==0.9.0
Werkzeug==0.14.1
